lost a day on this
The quickstart code and the tone of the dialog agree that at this point we are using data-driven logins, but this implementation in the documentation assumes hard-wired users. As a side-bar note, the failure to call .AddAspNetIdentity<ApplicationUser>() is one of the reasons users (yours truly being one) run into the "sub claim is missing" problem. All these assertions subject to verification, but this is what I'm seeing. :)