Feature Addition:
Given an Extension Grant as an entry condition, there is a need for the IProfileService to have knowledge of the incoming request parameters.
At the moment the only way to pass that knowledge of incoming parameters was to make up a fake claim on the identity in the IExtenstionGrantValidator implementations.

Given that extension grants are custom, the need for all data to be known when the final claims are built should be allowed.