Quickstart native client detection heuristic based on client's redirect uri
Redirection to a native client uses a loading page instead of an http
redirect. Previously, determination of the type of the client was based on
whether it used PKCE. This changes it to be based on the redirect uri. If it
has http or https scheme, an http redirect is used. Otherwise, we assume it's
a native client and use a loading page. This is based on the fact that a
native client should use a private-use uri scheme.