including dot net command to create mvc project
I think it would be useful for a newcomer (who might be exploring dot net core and identityserver, both at the same time, for the first time), to have the commands readily available in the tutorial itself. 

So, I have added the commands, allowing the tutorial to be more self-contained and useful to a new comer.