Update ExternalController.cs
There is this major bug where, the scheme used was leading to a complete failure of external login. The thing is, the scheme is correctly used in previous quickstarts but was not updated here. 

So, applying that fix. now, with this update, the quickstart is fully functional for external login. Tested with EF with migrations and an online azure sql server database.