specify "ASP.NET Identity" as "ASP.NET Core Identity"
I landed on this helpdoc as I'm trying to understand the relationship between ASP.NET Core Identity and IdentityServer4. I noticed in this doc a lot of references to "ASP.NET Identity", I believe it's being used as shorthand for "ASP.NET Core Identity". This PR specifies the term as "ASP.NET Core Identity" to remove the ambiguity.