Adding note about trusting certificates.
The quickstart uses `https` instead of `http` but the client in the quickstart will fail if the developer has not trusted the development certificate. Visual Studio is smart enough to prompt you to trust the localhost cert, but VS Code (which is what I was using) does not. I am proposing adding a note to explain this step. (also thanks to Daniel Tabuenca who helped me through this!)